update docs and create tutorial that corresponds to docs
