+++ examples/tutorial/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "tutorial"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+hewn = { path = "../../" }
